requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
aiosqlite>=0.20.0
numpy>=1.26.0
swisseph>=2.10.0
openai>=1.12.0
google-genai>=1.0.0
python-multipart>=0.0.9
Pillow>=10.0.0
razorpay>=1.4.1
stripe>=8.0.0
fpdf2>=2.7.0
slowapi>=0.1.9
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-httpx>=0.30.0
